Decode multiple frames if present in the file
A file or stream may consist of more than one frame. Do not stop
after the first frame when decoding in zstd. Continue reading until
we reach the end of the file or we get an error.

Move percent-done output into do_something() to avoid duplicating
code. Calculate the percentage as the ratio of bytes-in instead of
bytes-out since we won't know the total bytes-out until we process the
last frame.

Report the number of frames as well as the number of bytes at the end.
Report total valid checksums at the end instead of per-frame.